requirements.txt
beautifulsoup4==4.12.2
Flask==2.3.3
lxml==4.9.3
python-dotenv==1.0.0
requests==2.31.0